implemented live scholarship data

+++ scholarship/src/main/java/com/tesda8/region8/scholarship/model/dto/QualificationMapDto.java
@@ -3,13 +3,15 @@ package com.tesda8.region8.scholarship.model.dto;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 public class QualificationMapDto {
-    private Long amount;
+    private BigDecimal amount;
     private Long slots;
 
-    public QualificationMapDto(Long amount, Long slots) {
+    public QualificationMapDto(BigDecimal amount, Long slots) {
         this.amount = amount;
         this.slots = slots;
     }
